Fixed some serious bugs auto-adding and removing taxa to check lists.

+++ app/models/check_list.rb
@@ -78,7 +78,7 @@ class CheckList < List
   end
   
   def last_observation_of(taxon)
-    Observation.in_place(place).has_quality_grade(Observation::RESEARCH_GRADE).latest.first
+    Observation.of(taxon).in_place(place).has_quality_grade(Observation::RESEARCH_GRADE).latest.first
   end
   
   def self.sync_check_lists_with_parents(options = {})
@@ -128,11 +128,18 @@ class CheckList < List
     
     # get places places where this obs was made that already have this taxon listed
     existing_place_ids = ListedTaxon.all(:select => "id, list_id, place_id, taxon_id", 
-      :conditions => ["place_id IN (?) AND taxon_id IN (?)", place_ids, taxon_ids]).map(&:place_id)
+      :conditions => ["place_id IN (?) AND taxon_id = ?", place_ids, taxon_id]).map(&:place_id)
+    if observation
+      existing_place_ids += ListedTaxon.all(:select => "id, list_id, place_id, taxon_id", 
+        :conditions => ["place_id IN (?) AND last_observation_id = ?", place_ids, observation.id]).map(&:place_id)
+    end
       
     # if we need to add / update
     if observation && observation.quality_grade == Observation::RESEARCH_GRADE
-      ListedTaxon.update_all(["last_observation_id = ?", observation], ["place_id IN (?)", existing_place_ids])
+      ListedTaxon.update_all(
+        ["last_observation_id = ?", observation], 
+        ["place_id IN (?) AND taxon_id IN (?)", existing_place_ids, taxon_ids]
+      )
       return unless observation.taxon.rank == Taxon::SPECIES
       new_place_ids = place_ids - existing_place_ids
       CheckList.find_each(:conditions => ["place_id IN (?)", new_place_ids]) do |list|
@@ -141,9 +148,11 @@ class CheckList < List
     
     # otherwise well be refreshing / deleting
     else
-      ListedTaxon.find_each(:include => :list, :conditions => ["lists.type = 'CheckList' AND last_observation_id = ?", observation_id]) do |lt|
+      ListedTaxon.find_each(:include => :list, :conditions => [
+          "lists.type = 'CheckList' AND last_observation_id = ? OR list_id IN (?)", 
+          observation_id, existing_place_ids]) do |lt|
         obs = ListedTaxon.update_last_observation_for(lt)
-        if obs.blank? && !lt.user_id
+        if obs.blank? && !lt.user_id && !lt.updater_id && lt.comments_count.to_i == 0 && lt.list.is_default?
           lt.destroy
         end
       end
